# swagger.yaml
swagger: '2.0'
info:
  version: "1.0.0"
  title: Enfaixapp API
  description: |
    Aquesta API ha estat desenvolupada pel grup 13 de l'assignatura de [Projecte d'Enginyeria del Software](http://www.fib.upc.edu/fib/estudiar-enginyeria-informatica/assignatures/PES) de la [Facultat d'Informàtica de Barcelona](http://www.fib.upc.edu/fib.html) en el quadrimestre de Tardor 2016-2017 per a l'aplicació del món casteller Enfaixapp.
    
    Aquesta API només està disponible des de la xarxa d'adreces internes a la UPC o vía la seva VPN.    
    
    Desenvolupadors:
    - [Carlos Álamo](https://github.com/carlos2380)
    - [Àlex Correa](https://github.com/alexct33)
    - [Jordi Guiu](https://github.com/guiu23)
    - [Marc Martí](https://github.com/marc1161)
    - [Eduard Maura](https://github.com/mapu77)
    - [Arnau Orriols](https://github.com/aog182)
produces:
  - application/json
consumes:
  - application/json
host: 10.4.41.165:5000
...
paths:
...
  /colles:
    get:
      description: Obté totes les colles del sistema.El paràmetre opcional **type** permet filtrar el tipus de `Colla` castellera que es vol obtenir; o bé convencional (**conv**) o bé universitària (**uni**)
      tags:
        - Colles
      parameters:
        -
          name: type
          in: query
          description: Tipus de colla (**conv**encional o **uni**verstirària)
          required: false
          type: string
      responses:
        '200':
          description: Success
          schema:
            type: array
            items:
              $ref: '#/definitions/Colla'
        '400':
          description: Bad request
  /colles/{id}:
    get:
      description: Obté la `Colla` identificada pel paràmetre **id**
      tags:
        - Colles
      parameters:
        -
          name: id
          in: path
          description: Identificador únic de la colla
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: Success
          schema:
            $ref: "#/definitions/FullColla"
        '404':
          description: Resource not found
...
  /login:
    post:
      description: Inici de sessió
      tags:
        - Login
      parameters:
        -
          name: body
          in: body
          schema:
            $ref: '#/definitions/Login'
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/Usuari'
        '403':
          description: Unauthorized access
  /signin:
    post:
      description: Enregistrament
      tags:
      - Login
      parameters:
        -
           name: body
           in: body
           required: true
           schema:
            $ref: "#/definitions/NewUsuari"
      responses:
        '201':
          description: Resource created
          schema:
            $ref: "#/definitions/Usuari"
        '409':
          description: Conflict
...
definitions:
  FullColla:
    type: object
    description: Representa una colla castellera
    properties:
      id:
        description: Identificador únic de la colla
        type: integer
        format: int64
      name:
        description: Nom de la colla
        type: string
      uni:
        description: Expressa si és una colla universitària
        type: boolean
      description:
        description: Text introductori de la colla
        type: string
      email:
        description: Correu electrònic de contacte
        type: string
      web:
        description: Direcció web de la colla
        type: string
      address:
        description: Direcció postal de la colla
        type: string
      color:
        description: Codi RGB del color característic de la colla
        type: string
      img:
        description: Logotip de la colla castellera
        type: string
        format: byte
  Colla:
    type: object
    description: Resum d'una colla castellera
    properties:
      id:
        description: Identificador únic de la colla
        type: integer
        format: int64
      name:
        description: Nom de la colla
        type: string
      uni:
        description: Expressa si és una colla universitària
        type: boolean
      color:
        description: Codi RGB del color característic de la colla
        type: string
      img:
        description: Logotip de la colla castellera
        type: string
        format: byte
...
  Login:
    type: object
    description: Informació mínima per iniciar sessió
    required:
      - email
      - password
    properties:
      email:
        description: Correu electrònic de l'usuari
        type: string
      password:
        description: Contrasenya encriptada de l'usuari
        type: string
        format: password
  Usuari:
    type: object
    description: Representa un usuari del sistema
    properties:
      id:
        description: Identificador únic de l'usuari
        type: integer
        format: int64
      email:
        description: Correu electònic de l'usuari
        type: string
      name:
        description: Nom del'usuari
        type: string
      surname:
        description: Cognom de l'usuari
        type: string
      password:
        description: Contrasenya encriptada de l'usuari
        type: string
        format: password
      belongs:
        description: Llistat de les colles les quals pertany l'usuari
        type: array
        items:
          $ref: '#/definitions/Colla'
      follows:
        description: Llistat de lescolles les quals segueix l'usuari
        type: array
        items:
          $ref: "#/definitions/Colla"
      session_token:
        description: Token identificatiu de la sessió de l'usuari
        type: string
      admin:
        description: Expressa si l'usuari és administrador
        type: boolean
  NewUsuari:
    type: object
    description: Representa un usuari del sistema
    required:
    - email
    - name
    - surname
    - password
    properties:
      email:
        description: Correu electònic de l'usuari
        type: string
      name:
        description: Nom del'usuari
        type: string
      surname:
        description: Cognom de l'usuari
        type: string
      password:
        description: Contrasenya encriptada de l'usuari
        type: string
        format: password
      belongs:
        description: Llistat de les colles les quals pertany l'usuari
        type: array
        items:
          $ref: '#/definitions/Colla'
      follows:
        description: Llistat de lescolles les quals segueix l'usuari
        type: array
        items:
          $ref: "#/definitions/Colla"
...